split need to be simplified

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: tzi-qi <tzi-qi@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/23 18:37:58 by tzi-qi            #+#    #+#             */
-/*   Updated: 2022/05/23 19:38:14 by tzi-qi           ###   ########.fr       */
+/*   Updated: 2022/05/27 15:24:42 by tzi-qi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (size > 0)
